Removing duplicate css type definition for forms.
